src/econdata/econdata/spiders/db_utils.py
import psycopg2
from dotenv import load_dotenv
import os

# Load environment variables from .env file
load_dotenv()

# Database connection parameters
DB_NAME = os.getenv('DB_NAME')
DB_USER = os.getenv('DB_USER')
DB_PASSWORD = os.getenv('DB_PASSWORD')
DB_HOST = os.getenv('DB_HOST')
DB_PORT = os.getenv('DB_PORT')

def get_valid_currencies():
    # Connect to the PostgreSQL database
    conn = psycopg2.connect(
        dbname=DB_NAME,
        user=DB_USER,
        password=DB_PASSWORD,
        host=DB_HOST,
        port=DB_PORT
    )

    # Create a cursor object
    cur = conn.cursor()

    # Retrieve the list of valid currency codes from the currencies table
    cur.execute("SELECT code FROM public.currencies;")
    valid_currencies = {row[0] for row in cur.fetchall()}

    # Close the cursor and connection
    cur.close()
    conn.close()

    return valid_currencies