pyproject.toml
[project]
name = "econcell"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pip>=25.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
